+++ vaedecon/configs/__init__.py
@@ -13,3 +13,4 @@ from .default_config import TrainingConfig
 from .default_config import EvaluationConfig
 from .default_config import GEPDatasetConfig
 from .default_config import TestSetConfig
+from .default_config import TrainingSetSCTTargetConfig